feat: check if the user is an owner of auction

+++ backend/notes/views.py
@@ -34,15 +34,18 @@ class PrimaryKeyNoteView(APIView):
         serializer = NoteSerializer(
             current_note, data=request.data, context={
                 'current_user': request.user})
-        if serializer.is_valid(raise_exception=True):
+        if serializer.is_valid(
+                raise_exception=True) and current_note.owner == request.user:
             serializer.save()
             return Response(serializer.data, status=status.HTTP_200_OK)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk):
         current_note = Note.objects.get(pk=pk)
-        current_note.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if current_note.owner == request.user:
+            current_note.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
 
 
 class ArchivedNotesView(APIView):
@@ -50,6 +53,6 @@ class ArchivedNotesView(APIView):
     permission_classes = [permissions.IsAuthenticated]
 
     def get(self, request):
-        notes = Note.objects.filter(is_archived=True)
+        notes = Note.objects.filter(owner=request.user, is_archived=True)
         serializer = NoteSerializer(notes, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
